+++ ReleaseNotes.md
@@ -5,6 +5,7 @@ F# niggles part 3
 * Exempt methods of F# generated types with `@` in the name from ` ParameterNamesShouldMatchOverriddenMethodRule`
 * Exempt match on union types from `AvoidSwitchStatementsRule`
 * Exempt F# generated types with `@` in the name from `VariableNamesShouldNotMatchFieldNamesRule` and `UseCorrectCasingRule`
+* Exempt generated abstract closure types from `AbstractTypesShouldNotHavePublicConstructorsRule`
 
 For previous releases, go here -- https://github.com/SteveGilham/Gendarme/blob/master/ReleaseNotes%20-%20Previously.md
 Since this is a release from a fork, [issues should be reported at my related repo](https://github.com/SteveGilham/altcode.fake/issues) that contains a Fake driver for the Gendarme tool, but noted as being against the forked Gendarme tool itself.